+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-## 0.2.2 (unreleased)
+## 0.3.0 (unreleased)
 
 - Added support for `halfvec` and `sparsevec` types
 - Added support for `bit` type to Ecto
 - Added `Pgvector.extensions/0` function
 - Added `l1_distance`, `hamming_distance`, and `jaccard_distance` functions for Ecto
+- Dropped support for Elixir < 1.12
 
 ## 0.2.1 (2023-09-25)